feat: add exceptions

+++ server/src/main/java/ru/practicum/service/impl/CategoryServiceImpl.java
@@ -3,12 +3,16 @@ package ru.practicum.service.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.CategoryDto;
 import ru.practicum.dto.NewCategoryDto;
+import ru.practicum.exceptions.ConflictException;
+import ru.practicum.exceptions.NotFoundException;
 import ru.practicum.model.Category;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 import ru.practicum.service.CategoryService;
 import ru.practicum.util.mapper.CategoryMapper;
 
@@ -23,9 +27,15 @@ public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
 
+    private final EventRepository eventRepository;
+
     @Override
     @Transactional
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
+        if (categoryRepository.findAll().stream()
+                .anyMatch((category) -> category.getName().equals(newCategoryDto.getName()))) {
+            throw new ConflictException("Category with this name already exist");
+        }
         Category category = categoryRepository.save(CategoryMapper.toCategory(newCategoryDto));
         log.info("Category with id = {} saved", category.getId());
         return CategoryMapper.toCategoryDto(category);
@@ -34,6 +44,12 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public void deleteCategory(long catId) {
+        if (categoryRepository.findById(catId).isEmpty()) {
+            throw new NotFoundException("Category not found");
+        }
+        if (!eventRepository.findByCategory(catId, Pageable.ofSize(1)).isEmpty()) {
+            throw new ConflictException("Category associated with the event");
+        }
         categoryRepository.deleteById(catId);
         log.info("Category with id = {} deleted", catId);
     }
@@ -41,7 +57,12 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public CategoryDto patchCategory(long catId, CategoryDto categoryDto) {
-        Category category = categoryRepository.findById(catId).orElseThrow(); //todo: add exception
+        if (categoryRepository.findAll().stream()
+                .anyMatch((category) -> category.getName().equals(categoryDto.getName()))) {
+            throw new ConflictException("Category with this name already exist");
+        }
+        Category category = categoryRepository.findById(catId).orElseThrow(() ->
+                new NotFoundException(String.format("Category with id = %d found", catId)));
         category.setName(categoryDto.getName());
         category.setId(catId);
         Category newCategory = categoryRepository.save(category);
@@ -60,7 +81,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto getCategory(long catId) {
-        Category category = categoryRepository.findById(catId).orElseThrow();
+        Category category = categoryRepository.findById(catId)
+                .orElseThrow(() -> new NotFoundException(String.format("Category with id = %d found", catId)));
         log.info("Category with id = {} found", catId);
         return CategoryMapper.toCategoryDto(category);
     }
