+++ server/src/main/java/ru/practicum/service/impl/UserServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.UserDto;
 import ru.practicum.dto.request.NewUserRequest;
+import ru.practicum.exceptions.ConflictException;
+import ru.practicum.exceptions.NotFoundException;
 import ru.practicum.model.User;
 import ru.practicum.repository.UserRepository;
 import ru.practicum.service.UserService;
@@ -25,16 +27,25 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<UserDto> getUsers(List<Long> ids, int from, int size) {
-        List<UserDto> users = userRepository.findAll(PageRequest.of(from / size, size)).stream()
+        List<User> users;
+        if (ids.isEmpty()) {
+            users = userRepository.findAll(PageRequest.of(from / size, size))
+                    .stream().collect(Collectors.toList());
+        } else {
+            users = userRepository.findAllById(ids, PageRequest.of(from / size, size));
+        }
+        log.info("Users from {} to {} found", from, from + size);
+        return users.stream()
                 .map(UserMapper::toUserDto)
                 .collect(Collectors.toList());
-        log.info("Users from {} to {} found", from, from + size);
-        return users;
     }
 
     @Override
     @Transactional
     public UserDto createUser(NewUserRequest user) {
+        if (userRepository.findAll().stream().anyMatch(user1 -> user1.getEmail().equals(user.getEmail()))) {
+            throw new ConflictException("User with this email already exists");
+        }
         User userDto = userRepository.save(UserMapper.toUser(user));
         log.info("User with id = {} saved", userDto.getId());
         return UserMapper.toUserDto(userDto);
@@ -42,6 +53,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(long userId) {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NotFoundException("User not found");
+        }
         userRepository.deleteById(userId);
         log.info("User with id = {} deleted", userId);
     }
